Add corsheaders to settings.py and make Token names in headers capitalized, because the client does that too

+++ playlist/views.py
@@ -5,12 +5,25 @@ from rest_framework import status
 from .models import Playlist
 from .serializers import PlaylistSerializer
 from rest_framework.parsers import MultiPartParser, FormParser
+from game_admin.authentication import User
 import os
 
 class PlaylistPostView(APIView):
     '''Post view for Playlist model'''
     parser_classes = (MultiPartParser, FormParser)
-    def post(self, request):        
+    def post(self, request): 
+        # Get token1 and token2 from the request headers
+        # If the tokens are not valid, return a access denied response
+        token1 = request.headers.get('Token1')
+        token2 = request.headers.get('Token2')
+        
+        user = User()        
+
+        if not user.is_authorized(token1, token2):
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        
+        print("request.data", request.data) 
+        
         '''Post method for Playlist model'''
         playlist_serializer = PlaylistSerializer(data=request.data)
         if playlist_serializer.is_valid():
@@ -32,7 +45,14 @@ class PlaylistDeleteCategoryView(APIView):
     '''Delete view for Playlist model'''
     '''Delete all items from playlist that are in the category that has been passed in the url'''
     def delete(self, request, category):
-        '''Delete method for Playlist model'''
+         # Get token1 and token2 from the request headers
+        # If the tokens are not valid, return a access denied response
+        token1 = request.headers.get('token1')
+        token2 = request.headers.get('token2')
+        user = User()
+        if not user.is_authorized(token1, token2):
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        '''Delete category from the Playlist'''
         playlist = Playlist.objects.filter(category=category)
         playlist.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
@@ -41,7 +61,14 @@ class PlaylistDeleteItemView(APIView):
     '''Delete view for Playlist model'''
     '''Delete the item from playlist that has been passed in the url'''
     def delete(self, request, id):
-        '''Delete method for Playlist model'''
+         # Get token1 and token2 from the request headers
+        # If the tokens are not valid, return a access denied response
+        token1 = request.headers.get('token1')
+        token2 = request.headers.get('token2')
+        user = User()
+        if not user.is_authorized(token1, token2):
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        '''Delete item from the Playlist'''        
         playlist = Playlist.objects.filter(id=id)
         playlist.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
